backend/requirements.txt
fastapi
uvicorn
python-dotenv
motor
beanie
google-auth-oauthlib
google-api-python-client
google-generativeai
requests
httpx
apscheduler
python-multipart
python-jose[cryptography]
passlib[bcrypt]
pydantic-settings
pydantic[email]
